move some bigint stuff into mina-signer
